feat: move __get_mac function to ip_tools.py
